Rewrote files
Completely changed the implementation of Point in point.rs to restrict the type to only represent points in 3D space with three f32 members. This removed a lot of functionality. Rewrote line.rs and main.rs to reflect these changes. The original implementation of Point has been renamed to point_deprecated.rs and also includes the tests performed on this implementation from main.rs. This implementation will be saved as a reference for a future implementation of Point that will serve as a basis for a linear algebra API. Also began explicit lifetime handling for types using references to other types.